% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtrendshealth-package.R
\docType{package}
\name{gtrendshealth-package}
\alias{gtrendshealth}
\alias{gtrendshealth-package}
\title{gtrendshealth: Query the 'Google Trends for Health' API}
\description{
Connects to the 'Google Trends for Health' API hosted at \url{https://trends.google.com/trends/}, allowing projects authorized to use the health research data to query 'Google Trends'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CDCgov/gtrendshealth}
  \item Report bugs at \url{https://github.com/CDCgov/gtrendshealth/issues}
}

}
\author{
\strong{Maintainer}: Oscar de Leon \email{odeleon@alumni.emory.edu} (\href{https://orcid.org/0000-0003-1344-4412}{ORCID})

Authors:
\itemize{
  \item Oscar de Leon \email{odeleon@alumni.emory.edu} (\href{https://orcid.org/0000-0003-1344-4412}{ORCID})
}

Other contributors:
\itemize{
  \item US Centers for Disease Control and Prevention [copyright holder]
}

}
\keyword{internal}
