% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_gt_api_key}
\alias{remove_gt_api_key}
\title{Delete a saved GOOGLE TRENDS FOR HEALTH API Key}
\usage{
remove_gt_api_key(remove = FALSE, path = NULL)
}
\arguments{
\item{remove}{Whether to remove the key.}

\item{path}{Path to look for an .Renviron file.}
}
\description{
This function will uninstall your GOOGLE TRENDS FOR HEALTH API
key from the environment variables. If a path is provided, it will
be used to remove the key from .Renviron file in that path.
}
\keyword{internal}
